Add planetary and spacecraft data to context
