--- src/structure/integrity.rs
@@ -1,22 +1,0 @@
-/*
- * ANISE Toolkit
- * Copyright (C) 2021-2022 Christopher Rabotin <christopher.rabotin@gmail.com> et al. (cf. AUTHORS.md)
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/.
- *
- * Documentation: https://nyxspace.com/
- */
-
-use log::error;
-
-use crate::{
-    errors::{AniseError, IntegrityErrorKind},
-    structure::dataset::AniseContext,
-};
-
-impl<'a> AniseContext<'a> {
-    pub fn check_integrity(&self) -> Result<(), AniseError> {
-        Ok(())
-    }
-}
